+++ e2e-tests/custom-formats/CHANGELOG.md
@@ -1,5 +1,13 @@
 # custom-formats
 
+## 1.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @beff/client@0.0.135
+  - @beff/cli@0.0.135
+
 ## 1.0.2
 
 ### Patch Changes